# improve _fix_search_duplicates performance