@ultrabug 2021-12-09
Hello @AngryMane ; first of all, thank you very much for your contribution, it looks great!

I'm in the process of rewriting a part of the core to support #69 which I hope I should be able to finish next week

When I'm done, I'll be happy to merge your PR :+1: 

Please bear with me